refactor: 온도 관련 프로퍼티를 Double 타입으로 통일

+++ WeatherDress/Domain/Model/DailyWeather.swift
@@ -11,6 +11,6 @@ struct DailyWeather: Hashable {
     let date: Date
     let weatherCondition: WeatherCondition
     let rainfallProbability: Int
-    let maximumTemperature: Int
-    let minimunTemperature: Int
+    let maximumTemperature: Double
+    let minimunTemperature: Double
 }
